% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airline}
\alias{airline}
\title{Employee airline travel preferences survey data, UC Davis GreenFLY project}
\format{

}
\source{
Amenta, Annamaria; Sanguinetti, Angela (2019), Employee airline travel preferences survey data, UC Davis GreenFLY project, Dryad, Dataset, https://doi.org/10.25338/B81S5M
}
\usage{
airline
}
\description{
This survey data is from a study exploring the potential to promote lower-emissions air travel by providing consumers with information about the carbon emissions of possible flight choices in the context of online flight search and booking. Authors surveyed over 450 faculty, researchers, and staff at the University of California, Davis, and asked them to choose among hypothetical flight options for a domestic and an international university-related business trip. Emissions estimates for different flight alternatives Were displayed as prominently as price; this simple intervention has been promoted in several demonstration projects, including GreenFLY, a demo authors created at UC Davis.
}
\keyword{datasets}
