% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{awl}
\alias{awl}
\title{Academic Word List}
\format{
An object of class \code{data.frame} with 570 rows and 3 columns.
}
\source{
Coxhead, Averil (2000) A New Academic Word List. TESOL Quarterly, 34(2): 213-238.
}
\usage{
awl
}
\description{
The Academic Word List (AWL) was developed by Averil Coxhead as her MA thesis at the School of Linguistics and Applied Language Studies at Victoria University of wellington, New Zealand. The list contains 570 word families which were selected according to principles. The list does not include words that are in the most frequent 2000 words of English. The AWL was primarily made so that it could be used by teachers as part of a programme preparing learners for tertiary level study or used by students working alone to learn the words most needed to study at tertiary institutions. The Academic Word List replaces the University Word List.
}
\keyword{datasets}
