% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census2016}
\alias{census2016}
\alias{census_2016_G01}
\alias{census_2016_G04}
\alias{census_2016_G17}
\alias{census_2016_STE}
\title{Australian Bureau of Statistics 2016 Census Data}
\format{
An object of class \code{data.frame} with 9 rows and 109 columns.

An object of class \code{data.frame} with 9 rows and 307 columns.

An object of class \code{data.frame} with 9 rows and 481 columns.

An object of class \code{data.frame} with 9 rows and 3 columns.
}
\source{
Based on Australian Bureau of Statistics 2016 Census Data
}
\usage{
census_2016_G01

census_2016_G04

census_2016_G17

census_2016_STE
}
\description{
All the numerical values are counts according to the category given in the header and state or territory given as code `STE_CODE_2016` in row.
The data `census_2016_STE` has the code translation to the state or territory name.

Table G01 has counts of selected person characteristics by sex.

Table G04 has counts of poeple by age and sex.

Table G17 has counts of people by total weekly personal income by sex and age.
}
\keyword{datasets}
