% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwexercise-package.R
\docType{package}
\name{dwexercise-package}
\alias{dwexercise}
\alias{dwexercise-package}
\title{dwexercise: This package provides exercises for the data wrangling with R workshop}
\description{
Install this package to launch the `learnr` exercises for the course.
}
\keyword{internal}
