% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{frogsignal}
\alias{frogsignal}
\title{Multivariate phenotypic selection on sexual signal of frogs}
\format{
A data frame containing 535 rows and 22 variables.
\describe{
  \item{FrogID}{Individuals are identified with a 5-digit code. The first two digits indicate the year in which that female was collected and tested, while the next three digits are a unique identifier. }
  \item{AlternativeCD, AlternativeCR, AlternativeDF, AlternativeRA, AlternativePR}{We manipulated five call traits, and each female was given an alternative stimulus with a unique combination of these five traits. The trait values specific to each female are given in columns 2-6: “AlternativeCD” (call duration), ”AlternativeCR” (call rate), “AlternativeDF” (dominant frequency), “AlternativeRA” (relative amplitude), and “AlternativePR” (pulse rate).}
}
}
\source{
Tanner, Jessie C.; Ward, Jessica L.; Shaw, Ruth G.; Bee, Mark A. (2017), Data from: Multivariate phenotypic selection on a complex sexual signal, Dryad, Dataset, \url{https://doi.org/10.5061/dryad.n50cr}
}
\usage{
frogsignal
}
\description{
A dataset containing infomraiton on receivers.
}
\keyword{datasets}
