% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oz_climate}
\alias{oz_climate}
\alias{oz_climate_qbook}
\title{Australia's attitudes towards climate change}
\format{
An object of class \code{data.frame} with 1927 rows and 200 columns.

An object of class \code{data.frame} with 200 rows and 2 columns.
}
\source{
Morrison, Mark; Duncan, Roderick; Parton, Kevin (2015), Data from: Religion does matter for climate change attitudes and behavior, Dryad, Dataset, https://doi.org/10.5061/dryad.vr315
}
\usage{
oz_climate

oz_climate_qbook
}
\description{
Using an online survey of 1,927 Australians, authors examined links between membership of four religious groupings (Buddhists, Christian literalists and non-literalists, and Secularists) and climate change attitudes and behaviors.

`oz_climate_qbook` has the translation of the column label to the actual question asked. If the value is missing, the authors did not have the labels in the original data set.
}
\keyword{datasets}
