% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tidy_abstracts}
\alias{tidy_abstracts}
\title{Abstracts for various papers associated with tidyverse}
\format{
An object of class \code{data.frame} with 8 rows and 5 columns.
}
\usage{
tidy_abstracts
}
\description{
The data contains 8 articles that are related to `tidyverse` that contains authors, year it was published, which journal it was published in, the title and abstract.
}
\keyword{datasets}
