% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{treefall}
\alias{treefall}
\title{Empirical relationships between tree fall and landscape-level amounts of logging and fire}
\format{
An object of class \code{data.frame} with 737 rows and 12 columns.
}
\source{
Lindenmayer, David B. et al. (2019), Data from: Empirical relationships between tree fall and landscape-level amounts of logging and fire, Dryad, Dataset, \url{https://doi.org/10.5061/dryad.h70j3}
}
\usage{
treefall
}
\description{
Empirical relationships between tree fall and landscape-level amounts of logging and fire
}
\keyword{datasets}
